1. 增加系统平台自适应模块 multiplatforms.py 2. 修改 HPSocket.py 以适配 Linux-HPSocket-1.0.0 3. 修复了 test_server.py 中在 Linux 下的一个报错

+++ test_server.py
@@ -81,7 +81,7 @@ class HP_TcpPullServer():
 
         iAddressLen = 50
         pszAddress  = ctypes.create_string_buffer(b' ' * iAddressLen)
-        iAddressLen = ctypes.c_long(iAddressLen)
+        iAddressLen = ctypes.c_int(iAddressLen)
         usPort      = ctypes.c_ushort(0)
         HPSocket.HP_Server_GetListenAddress(Sender, pszAddress, ctypes.byref(iAddressLen), ctypes.byref(usPort))
 
@@ -94,7 +94,7 @@ class HP_TcpPullServer():
 
         iAddressLen  = 50
         pszAddress   = ctypes.create_string_buffer(b' ' * iAddressLen)  # 这里要预留空间，GetRemoteAddress的调用方负责管理内存
-        iAddressLen  = ctypes.c_long(iAddressLen)
+        iAddressLen  = ctypes.c_int(iAddressLen)
         usPort       = ctypes.c_ushort(0)
         HPSocket.HP_Server_GetRemoteAddress(Sender, ConnID, pszAddress, ctypes.byref(iAddressLen), ctypes.byref(usPort))
